Time: 26 ms (25.72%) | Memory: 45.3 MB (34.63%) - LeetSync

+++ LC/907-koko-eating-bananas/koko-eating-bananas.java
@@ -1,8 +1,8 @@
 class Solution {
-    public static int findHours (int[] arr, double speed) {
+    public static int findHours (int[] arr, int speed) {
         int time = 0;
         for (int i = 0; i < arr.length; i++) {
-            time += Math.ceil(arr[i] / speed);
+            time += Math.ceil((double) arr[i] / speed);
         }
         return time;
     }
